cambiado a float porque no funcionaba decimal

+++ database/migrations/2022_06_28_182723_create_puntos_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->timestamps();
             $table->string('NameLocal');
             $table->Text('Description');
-            $table->decimal('Longitud');
-            $table->decimal('Latitud');
+            $table->float('Longitud');
+            $table->float('Latitud');
             $table->integer('Type');
             $table->string('User');
         });
